Adapter 파일 분리
