+++ app/src/main/java/campus/tech/kakao/map/Place.kt
@@ -1,0 +1,7 @@
+package campus.tech.kakao.map
+
+data class Place(
+    val name: String,
+    val address: String,
+    val category: String
+)
